cambia la fuente de las imagenes y agrega badges en vectores
