+++ index.html
@@ -9,8 +9,8 @@
   <title>Pokedex</title>
 
   <link href="node_modules/bootstrap/dist/css/bootstrap.min.css" rel="stylesheet">
-  <link rel="stylesheet" href="node_modules/@tarekraafat/autocomplete.js/dist/css/autoComplete.css">
-  <link href="src/css/style.css" rel="stylesheet">
+  <link rel="stylesheet" href="./src/css/autoComplete.css">
+  <link rel="stylesheet" href="./src/css/style.css">
 
 </head>
 
@@ -19,8 +19,10 @@
     <h1 class="display-3">Encontrá tu pokemon</h1>
 
     <hr class="my-2">
-
-    <input id="autoComplete" class="form-control mr-sm-2" type="search" placeholder="Buscar" aria-label="Buscar">
+    <div class="container">
+      <input id="autoComplete" class="form-control mr-sm-2" type="search" placeholder="Buscar" aria-label="Buscar">
+    </div>
+    
   </div>
   <div class="d-flex" id="wrapper">
