+++ src/js/modules/ui.js
@@ -2,6 +2,27 @@ const $listaContainer = $('.container-lista');
 const urlPokemon = 'https://pokeapi.co/api/v2/pokemon';
 
 function mostrarPokemonSeleccionado(pokemonSeleccionado) {
+	const pokemonTypes = {
+		path:'../img/type-badges/',
+		bug:'bug.svg',
+		dark:'dark.svg',
+		dragon:'dragon.svg',
+		electric:'',
+		fairy:'',
+		fighting:'',
+		fire:'',
+		flying:'',
+		ghost:'',
+		grass:'',
+		ground:'',
+		ice:'',
+		normal:'',
+		poison:'',
+		psychic:'',
+		rock:'',
+		steel:'',
+		water:''
+	};
 	const containerPokemon = $('<div id="container-pokemon" class="card mx-auto mt-5"></div>');
 	const nombrePokemon = $('<h2 id="nombre-pokemon" class="card-title pt-3 text-capitalize text-center">Cargando...</h2>');
 	const containerBadgeTipoPokemon = $('<div id="container-badge-tipo-pokemon" class="container"></div>');
@@ -23,7 +44,7 @@ function mostrarPokemonSeleccionado(pokemonSeleccionado) {
 	containerPokemon.append(containerBadgeTipoPokemon);
   
 	for (let i = 0; i < pokemonSeleccionado.tipos.length; i ++) {
-	  containerBadgeTipoPokemon.append($(`<span class="badge pb-1 ${pokemonSeleccionado.tipos[i]}">${pokemonSeleccionado.tipos[i]}</span>`));
+	  containerBadgeTipoPokemon.append($(`<img class="icon ${pokemonSeleccionado.tipos[i]}" src="./src/img/pokemon-types/${pokemonSeleccionado.tipos[i]}.svg"/>`));
 	}
   
 	containerPokemon.append(fotoPokemon);
